+++ changelog.md
@@ -3,6 +3,7 @@
 ## Version 2.3.1 (Future)
 - Allow dropdown actions to present as split buttons
 - Make param varchar on permissions table
+- Add route for displaying bozboz package versions
 
 ## Verison 2.3.0 (2017-01-04)
 - Lock down user/role editing so sneaky clients can't assign themselves as Bozboz admin